EOMC: moved eomc parser inside namespace eomc
EOMC: renamed utils and parser file following icub-main naming convention
