+++ src/libraries/icubmod/embObjMotionControl/eomcUtils.h
